remove stripped wood post
